fix(bundle): installation of bundle on openshift
This commit fixes the installation of bundle on openshift which fails
because of the required CRD servicemonitor (which is provided by
Prometheus Operator). Although the CRD is present in the cluster, OLM
seems to be installing PO which conflicts with the in-cluster Prometheus
Operator. The fix is to remove the dependency in CSV on ServiceMonitors

Signed-off-by: Sunil Thaha <sthaha@redhat.com>