fix(bundle): update bundle only if there are changes
Previously, `make bundle` will update the `createdAt` of the bundle CSV
file even if there aren't any changes to the config. This requires PRs
to be rebased to resolve the conflict in createdAt. This PR changes the
bundle generation to revert the CSV file if the only change detected is
for the `createdAt` timestamp, there by saving us from unnecessary
rebases.

Signed-off-by: Sunil Thaha <sthaha@redhat.com>