ci: test bundle-upgrade scenario
Previously, in CI, the operator bundle was installed on a clean cluster
instead of the scenario where the bundle is upgraded. This has resulted
in bundle getting rejected in Community Operator Catalog when the
upgrade test failed.

This commit modifies the e2e test to run the upgrade scenario instead of
installing on a fresh cluster.

Signed-off-by: Sunil Thaha <sthaha@redhat.com>