fix(test): provide more time for operator to initialize
This commit increases the liveness and the readiness probes' timeout so
that the operator has time to build its internal cache and be ready
for reconcile.

Tests fail because the resources (kepler) are recreated and expected to
be reconciled before the operator has had a change to build cache and
the informers.

Signed-off-by: Sunil Thaha <sthaha@redhat.com>