feat(webhook): add webhook to validate kepler resource
Previously, when user created kepler with name other `kepler`, k8s API
server would accept the resource and create it. The operator then
updates the status of the resource to be `invalid`. This is now changed
by adding Validating webhook that rejects the resource if the name does
not match `kepler`.

Signed-off-by: Sunil Thaha <sthaha@redhat.com>