chore(tests): add flag to accept kepler image for internal tests
This commit adds a flag to accept Kepler image which will be used when
running Kepler Internal tests. If there is no value provided then it
will use the default Kepler image. It is useful when running the tests
in downstream CI where the Kepler images are not pulled from quay and
are built from pipelines

Signed-off-by: vprashar2929 <vibhu.sharma2929@gmail.com>