chore: make kepler CR deprecated
This commit sets the Kepler CR status as deprecated. It also removes the
existing reconcilers ensuring that when Kepler CR is created it only
reconciles openshift dashboards and sets the status of CR as degraded

It also adds a new conditions field to Kepler status which ensures that
deprecation status is reflected properly on OpenShift console.

Signed-off-by: vprashar2929 <vibhu.sharma2929@gmail.com>