chore: use build-manager target for building operator
This commit adds a new target to the Makefile to build the operator.
`make build-manager` will only be responsible for building the operator
`make build` will run the dependencies and build the operator.

It also updates the Dockerfile to use the build-manager target and pass
on the required arguments to the build-manager.

Signed-off-by: Vibhu Prashar <vibhu.sharma2929@gmail.com>