fix: remove security mode from samples
This commit removes the security mode from the samples. The security
mode was set to none by default and when the operator was deployed
using default spec on OpenShift it was deployed with the security mode
set to none. Ideally we shouldn't have set this in samples so that
the operator can be deployed with the default security mode i.e rbac

Signed-off-by: vprashar2929 <vibhu.sharma2929@gmail.com>